Ensure the the DB seeds don't are indempotent
Meaning, when running multiple times they should not duplicate data.
Note that when creating Jane we can’t use `find_or_create_by`. That’s
because `password` is not actually a field on the model.

Finally, there was also an error when creating responses. We can’t
directly set `seat`. We have to set `seat_id`.