Truncate response bodies over 65,535 characters
That number seems pretty high, and we may want to reduce it at some
point. For now, I figured we’d see how it works out.

Note that this limit is not being enforced in the DB. Not sure why, but
I was not able to set a limit on the text column in postgres.