Use a flex-based layout for the sidebar and main area
Bootstrap columns work well, but are always responsive. Here, we want
the left side navigation to have a fixed width. Note that the flex
layout sass is in the /lib directory. This is because it theoretically
could be used in completely separate apps, and is not specific to this
app.