Don't send more than one summary email per team
Classic `joins` duplication. This is also a good illustration as to why
test assertions should be against hard-coded data. In this case, the
test data logic mirrored the actual logic. There was a bug in the
actual logic, and therefore a bug in our tests.

I’ve now hardcoded the count of teams with unhandled responses in the
tests. The downside if this is that if we add more fixtures, the test
will fail. I’m taking that trade-off for now, and hopefully in the
future I can come up with a way to meet both goals.